+++ x/zoneconcierge/keeper/grpc_query.go
@@ -6,7 +6,6 @@ import (
 	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
 	"github.com/babylonlabs-io/babylon/v3/x/zoneconcierge/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/query"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
@@ -20,104 +19,6 @@ func (k Keeper) Params(c context.Context, req *types.QueryParamsRequest) (*types
 	ctx := sdk.UnwrapSDKContext(c)
 
 	return &types.QueryParamsResponse{Params: k.GetParams(ctx)}, nil
-}
-
-// Header returns the header at a given height
-func (k Keeper) Header(c context.Context, req *types.QueryHeaderRequest) (*types.QueryHeaderResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "invalid request")
-	}
-
-	if len(req.ConsumerId) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "chain ID cannot be empty")
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-
-	header, err := k.GetHeader(ctx, req.ConsumerId, req.Height)
-	if err != nil {
-		return nil, err
-	}
-	resp := &types.QueryHeaderResponse{
-		Header: header,
-	}
-
-	return resp, nil
-}
-
-// EpochChainsInfo returns the latest info for list of chains in a given epoch
-func (k Keeper) EpochChainsInfo(c context.Context, req *types.QueryEpochChainsInfoRequest) (*types.QueryEpochChainsInfoResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "invalid request")
-	}
-
-	// return if no chain IDs are provided
-	if len(req.ConsumerIds) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "consumer IDs cannot be empty")
-	}
-
-	// return if chain IDs contain duplicates or empty strings
-	if err := bbntypes.CheckForDuplicatesAndEmptyStrings(req.ConsumerIds); err != nil {
-		return nil, status.Error(codes.InvalidArgument, types.ErrInvalidConsumerIDs.Wrap(err.Error()).Error())
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-	var chainsInfo []*types.ChainInfo
-	// TODO: paginate this for loop
-	for _, ConsumerId := range req.ConsumerIds {
-		// check if chain ID is valid
-		if !k.HasChainInfo(ctx, ConsumerId) {
-			return nil, status.Error(codes.InvalidArgument, types.ErrChainInfoNotFound.Wrapf("chain ID %s", ConsumerId).Error())
-		}
-
-		// if the chain info is not found in the given epoch, return with empty fields
-		if !k.EpochChainInfoExists(ctx, ConsumerId, req.EpochNum) {
-			chainsInfo = append(chainsInfo, &types.ChainInfo{ConsumerId: ConsumerId})
-			continue
-		}
-
-		// find the chain info of the given epoch
-		chainInfoWithProof, err := k.GetEpochChainInfo(ctx, ConsumerId, req.EpochNum)
-		if err != nil {
-			return nil, err
-		}
-
-		chainsInfo = append(chainsInfo, chainInfoWithProof.ChainInfo)
-	}
-
-	resp := &types.QueryEpochChainsInfoResponse{ChainsInfo: chainsInfo}
-	return resp, nil
-}
-
-// ListHeaders returns all headers of a chain with given ID, with pagination support
-func (k Keeper) ListHeaders(c context.Context, req *types.QueryListHeadersRequest) (*types.QueryListHeadersResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "invalid request")
-	}
-
-	if len(req.ConsumerId) == 0 {
-		return nil, status.Error(codes.InvalidArgument, "chain ID cannot be empty")
-	}
-
-	ctx := sdk.UnwrapSDKContext(c)
-
-	headers := []*types.IndexedHeader{}
-	store := k.canonicalChainStore(ctx, req.ConsumerId)
-	pageRes, err := query.Paginate(store, req.Pagination, func(key, value []byte) error {
-		var header types.IndexedHeader
-		k.cdc.MustUnmarshal(value, &header)
-		headers = append(headers, &header)
-		return nil
-	})
-	if err != nil {
-		return nil, status.Error(codes.Internal, err.Error())
-	}
-
-	resp := &types.QueryListHeadersResponse{
-		Headers:    headers,
-		Pagination: pageRes,
-	}
-	return resp, nil
 }
 
 // FinalizedChainsInfo returns the finalized info for a given list of chains
@@ -237,8 +138,6 @@ func (k Keeper) FinalizedChainInfoUntilHeight(c context.Context, req *types.Quer
 		finalizedEpoch = chainInfo.LatestHeader.BabylonEpoch
 	}
 
-	resp.FinalizedChainInfo = chainInfo
-
 	if chainInfo.LatestHeader.Height <= req.Height { // the requested height is after the last finalised chain info
 		// find and assign the epoch metadata of the finalised epoch
 		resp.EpochInfo, err = k.epochingKeeper.GetHistoricalEpoch(ctx, finalizedEpoch)
@@ -267,11 +166,6 @@ func (k Keeper) FinalizedChainInfoUntilHeight(c context.Context, req *types.Quer
 		}
 		// assign the finalizedEpoch, and retrieve epoch info, raw ckpt and submission key
 		finalizedEpoch = closestHeader.BabylonEpoch
-		chainInfoWithProof, err := k.GetEpochChainInfo(ctx, req.ConsumerId, finalizedEpoch)
-		if err != nil {
-			return nil, err
-		}
-		resp.FinalizedChainInfo = chainInfoWithProof.ChainInfo
 		resp.EpochInfo, err = k.epochingKeeper.GetHistoricalEpoch(ctx, finalizedEpoch)
 		if err != nil {
 			return nil, err