+++ x/zoneconcierge/keeper/grpc_query_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/babylonlabs-io/babylon/v3/app"
 	btclightclienttypes "github.com/babylonlabs-io/babylon/v3/x/btclightclient/types"
-	"github.com/cosmos/cosmos-sdk/types/query"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 
@@ -21,33 +20,6 @@ type chainInfo struct {
 	consumerID        string
 	numHeaders        uint64
 	headerStartHeight uint64
-}
-
-func FuzzHeader(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-
-		babylonApp := app.Setup(t, false)
-		zcKeeper := babylonApp.ZoneConciergeKeeper
-		ctx := babylonApp.NewContext(false)
-
-		// invoke the hook a random number of times to simulate a random number of blocks
-		numHeaders := datagen.RandomInt(r, 100) + 2
-		headers := SimulateNewHeaders(ctx, r, &zcKeeper, consumerID, 0, numHeaders)
-
-		// find header at a random height and assert correctness against the expected header
-		randomHeight := datagen.RandomInt(r, int(numHeaders-1))
-		resp, err := zcKeeper.Header(ctx, &zctypes.QueryHeaderRequest{ConsumerId: consumerID, Height: randomHeight})
-		require.NoError(t, err)
-		require.Equal(t, headers[randomHeight].Header.AppHash, resp.Header.Hash)
-
-		// find the last header and assert correctness
-		resp, err = zcKeeper.Header(ctx, &zctypes.QueryHeaderRequest{ConsumerId: consumerID, Height: numHeaders - 1})
-		require.NoError(t, err)
-		require.Equal(t, headers[numHeaders-1].Header.AppHash, resp.Header.Hash)
-	})
 }
 
 func FuzzEpochChainsInfo(f *testing.F) {
@@ -103,73 +75,6 @@ func FuzzEpochChainsInfo(f *testing.F) {
 
 			// simulate the scenario that a random epoch has ended
 			hooks.AfterEpochEnds(ctx, epochNum)
-		}
-
-		// assert correctness of best case scenario
-		for _, epochNum := range epochNums {
-			resp, err := zcKeeper.EpochChainsInfo(ctx, &zctypes.QueryEpochChainsInfoRequest{EpochNum: epochNum, ConsumerIds: consumerIDs})
-			require.NoError(t, err)
-			epochChainsInfo := resp.ChainsInfo
-			require.Len(t, epochChainsInfo, int(numChains))
-			for _, info := range epochChainsInfo {
-				actualHeight := epochToChainInfo[epochNum][info.ConsumerId].headerStartHeight + (epochToChainInfo[epochNum][info.ConsumerId].numHeaders - 1)
-				require.Equal(t, actualHeight, info.LatestHeader.Height)
-			}
-		}
-
-		// if num of chain ids exceed the max limit, query should fail
-		largeNumChains := datagen.RandomInt(r, 10) + 101
-		var maxConsumerIDs []string
-		for i := uint64(0); i < largeNumChains; i++ {
-			maxConsumerIDs = append(maxConsumerIDs, datagen.GenRandomHexStr(r, 30))
-		}
-		randomEpochNum := datagen.RandomInt(r, 10) + 1
-		_, err := zcKeeper.EpochChainsInfo(ctx, &zctypes.QueryEpochChainsInfoRequest{EpochNum: randomEpochNum, ConsumerIds: maxConsumerIDs})
-		require.Error(t, err)
-
-		// if no input is passed in, query should fail
-		_, err = zcKeeper.EpochChainsInfo(ctx, &zctypes.QueryEpochChainsInfoRequest{EpochNum: randomEpochNum, ConsumerIds: nil})
-		require.Error(t, err)
-
-		// if len of chain ids is 0, query should fail
-		_, err = zcKeeper.EpochChainsInfo(ctx, &zctypes.QueryEpochChainsInfoRequest{EpochNum: randomEpochNum, ConsumerIds: []string{}})
-		require.Error(t, err)
-
-		// if chain ids contain duplicates, query should fail
-		randomConsumerID := datagen.GenRandomHexStr(r, 30)
-		dupConsumerIds := []string{randomConsumerID, randomConsumerID}
-		_, err = zcKeeper.EpochChainsInfo(ctx, &zctypes.QueryEpochChainsInfoRequest{EpochNum: randomEpochNum, ConsumerIds: dupConsumerIds})
-		require.Error(t, err)
-	})
-}
-
-func FuzzListHeaders(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-
-		babylonApp := app.Setup(t, false)
-		zcKeeper := babylonApp.ZoneConciergeKeeper
-		ctx := babylonApp.NewContext(false)
-
-		// invoke the hook a random number of times to simulate a random number of blocks
-		numHeaders := datagen.RandomInt(r, 100) + 1
-		headers := SimulateNewHeaders(ctx, r, &zcKeeper, consumerID, 0, numHeaders)
-
-		// a request with randomised pagination
-		limit := datagen.RandomInt(r, int(numHeaders)) + 1
-		req := &zctypes.QueryListHeadersRequest{
-			ConsumerId: consumerID,
-			Pagination: &query.PageRequest{
-				Limit: limit,
-			},
-		}
-		resp, err := zcKeeper.ListHeaders(ctx, req)
-		require.NoError(t, err)
-		require.Equal(t, int(limit), len(resp.Headers))
-		for i := uint64(0); i < limit; i++ {
-			require.Equal(t, headers[i].Header.AppHash, resp.Headers[i].Hash)
 		}
 	})
 }